refactor(OutputView): 통계 뷰 분리
